move and improve performance_evaluator
